ISSUE #4780 - refactoring
