+++ frontend/src/v5/validation/containerAndFederationSchemes/validators.ts
@@ -17,56 +17,17 @@
 
 import * as Yup from 'yup';
 import { formatMessage } from '@/v5/services/intl';
-import filesize from 'filesize';
-import { trimmedString } from '../shared/validators';
+import { alphaNumericHyphens, stripIfBlankString } from '../shared/validators';
 import { getState } from '@/v4/modules/store';
 import { selectRevisions, selectRevisionsPending } from '@/v5/store/containerRevisions/containerRevisions.selectors';
 import { ContainerRevisionsActionsDispatchers } from '@/v5/services/actionsDispatchers';
-
-const stripIfBlankString = (value) => (
-	value === ''
-		? Yup.string().strip()
-		: Yup.string());
-
-export const numberField = Yup.number()
-	.default(0)
-	.typeError(formatMessage({
-		id: 'settings.surveyPoint.error.number',
-		defaultMessage: 'Must be a number',
-	}));
 
 export const nullableNumberField = Yup.number()
 	.transform((value) => (Number.isNaN(value) ? null : value))
 	.nullable();
 
-export const name = trimmedString
-	.max(120,
-		formatMessage({
-			id: 'validation.model.name.error.max',
-			defaultMessage: 'Name is limited to 120 characters',
-		}))
-	.required(
-		formatMessage({
-			id: 'validation.model.name.error.required',
-			defaultMessage: 'Name is a required field',
-		}),
-	)
-	.test(
-		'alreadyExistingNames',
-		formatMessage({
-			id: 'validation.model.name.alreadyExisting',
-			defaultMessage: 'This name is already used within this project',
-		}),
-		(nameValue, testContext) => {
-			if (!testContext.options?.context) return true;
-			return !testContext.options.context.alreadyExistingNames?.map((n) => n.trim().toLocaleLowerCase()).includes(nameValue?.toLocaleLowerCase());
-		},
-	);
-
 export const unit = Yup.string().required().oneOf(['mm', 'cm', 'dm', 'm', 'ft']).default('mm');
 export const type = Yup.string().required().default('Uncategorised');
-
-export const alphaNumericHyphens = /^[\w|_|-]*$/;
 
 export const code = Yup.lazy((value) => (
 	stripIfBlankString(value)
@@ -79,15 +40,6 @@ export const code = Yup.lazy((value) => (
 			formatMessage({
 				id: 'validation.model.code.error.characters',
 				defaultMessage: 'Code can only consist of letters, numbers, hyphens or underscores',
-			}))
-));
-
-export const desc = Yup.lazy((value) => (
-	stripIfBlankString(value)
-		.max(660,
-			formatMessage({
-				id: 'validation.model.description.error.max',
-				defaultMessage: 'Description is limited to 660 characters',
 			}))
 ));
 
@@ -139,14 +91,5 @@ export const revisionDesc = Yup.lazy((value) => (
 				defaultMessage: 'Revision Description is limited to 660 characters',
 			}))
 ));
-
-export const uploadFile = Yup.mixed().nullable().test(
-	'fileSize',
-	formatMessage({
-		id: 'validation.revisions.file.error.tooLarge',
-		defaultMessage: 'File exceeds size limit of {sizeLimit}',
-	}, { sizeLimit: filesize(ClientConfig.uploadSizeLimit) }),
-	({ size }) => size && (size < ClientConfig.uploadSizeLimit),
-);
 
 export const lod = Yup.number();
